requirements.txt
disnake==2.1.0
emoji==1.7.0
ffmpeg==1.4
GitPython==3.1.30
google==3.0.0
googletrans==3.0.0
gTTS==2.2.4
jishaku==2.5.0
motor==3.0.0
numpy==1.23.1
pymongo==4.2.0
python-dotenv==0.20.0
pytube==12.1.2
pytz==2022.1
requests==2.28.1
tweepy==4.12.1
ping3==4.0.4

